Consume User Creation Webhooks From Clerk
This commit updates the webhook handler that responds to Clerk
webhooks. It does not add validation to the handler, only implements
the use case, etc.

This also fixes the copy on the README to me be more clear that
we are syncing WEBSOCKET MESSAGES into the Redux store and not WEBHOOKS